+++ .sequelizerc
@@ -1,0 +1,8 @@
+const path = require('path')
+
+module.exports = {
+  "config": path.resolve('./server/config', 'config.js'),
+  "models-path": path.resolve('./server/models'),
+  "seeders-path": path.resolve('./server/seeders'),
+  "migrations-path": path.resolve('./server/migrations')
+}
