+++ .vscode/launch.json
@@ -11,7 +11,7 @@
       "skipFiles": [
         "<node_internals>/**"
       ],
-      "program": "${workspaceFolder}\\app.js"
+      "program": "${workspaceFolder}\\index.js"
     }
   ]
 }
