--- config/index.js
@@ -1,10 +1,0 @@
-const dbConfig = require('./config')
-
-// set other config ......
-
-const value = {
-  dbConfig: dbConfig,
-  env: process.env.NODE_ENV || 'development'
-}
-
-module.exports = value
