+++ docker-compose.yaml
@@ -32,6 +32,8 @@ services:
       POSTGRES_USERNAME: postgres
     volumes:
       - postgres_data:/var/lib/postgresql/data
+    ports:
+      - "5432:5432"
 
 volumes:
   postgres_data: {}
