+++ server/routes/logout.js
